Added NPC2
No patrol route made. Though the working can be tested by just moving
around 'taichi'. The NPC 2 will follow and try to attack player. Here's
the major bug - When the NPC bends to attack he falls down below and the
animation is not completed. I'm not really sure y this happens. Maybe
it's the nav mesh or may be its the animation clip. However, it works
perfectly fine in my milestone 2 scene. Maybe you guys can identify the
cause of it!